requirements.txt
aenum==2.1.2
alembic==1.5.2
amqp==5.0.2
aniso8601==8.1.0
anyio==3.3.3
asgiref==3.3.1
asn1crypto==0.24.0
astor==0.7.1
attrdict==2.0.1
attrs==18.2.0
autobahn==20.7.1
Automat==0.7.0
autopep8==1.4.3
base58==2.0.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bip32utils==0.3.post4
bitarray==1.2.2
bitcoin==1.1.42
blessings==1.7
bpython==0.17.1
bs4==0.0.1
bumpversion==0.5.3
cachetools==4.2.4
celery==5.0.5
certifi==2018.11.29
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudscraper==1.2.48
colander==1.8.3
colorlog==4.0.2
constantly==15.1.0
construct==2.10.67
coverage==4.5.2
coveralls==1.5.1
coz-bytecode==0.5.1
cryptography==3.2.1
curtsies==0.3.0
cycler==0.10.0
cytoolz==0.11.0
dateparser==1.0.0
defusedxml==0.6.0
Django==3.2
django-allauth==0.51.0
django-appconf==1.0.5
django-celery-beat==2.2.0
django-celery-email==3.0.0
django-cors-headers==3.13.0
django-extensions==2.2.6
django-filter==2.4.0
django-rest-auth==0.9.1
django-timezone-field==4.1.1
djangorestframework==3.13.1
docopt==0.6.2
ecdsa==0.13
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
ethereum==2.0.4
Events==0.3
Flask==0.12.2
Flask-Ext==0.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
furl==2.0.0
gevent==1.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gql==2.0.0
graphql-core==2.3.2
greenlet==0.4.15
gunicorn==19.7.1
h11==0.12.0
hdwallet==1.3.0
hexbytes==0.2.1
httpcore==0.13.7
httplib2==0.18.1
httpx==0.20.0
hyperlink==18.0.0
idna==2.8
importlib-metadata==3.1.1
importlib-resources==3.3.0
incremental==17.5.0
ipfshttpclient==0.6.1
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
jwt==1.1.0
klein==17.10.0
kombu==5.0.2
libeospy==2.0.1
logzero==1.5.0
lru-dict==1.1.6
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
memory-profiler==0.55.0
mmh3==2.5.1
mnemonic==0.20
mock==2.0.0
mpmath==1.1.0
multiaddr==0.0.9
mypy-extensions==0.4.3
Naked==0.1.31
neo-boa==0.5.6
neo-python==0.8.4
neo-python-rpc==0.2.1
neocore==0.5.6
netaddr==0.8.0
newtulipy==0.4.2
numpy==1.21.4
oauthlib==3.1.0
orderedmultidict==1.0.1
parsimonious==0.8.1
pbkdf2==1.3
pbr==5.1.1
peewee==3.8.1
pexpect==4.6.0
pika==1.2.0
Pillow==8.0.1
pluggy==0.8.1
plyvel==1.0.5
promise==2.3
prompt-toolkit==2.0.7
protobuf==3.14.0
psutil==5.4.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyblake2==1.1.2
pycodestyle==2.4.0
pycparser==2.19
pycryptodome==3.7.2
pyethash==0.1.27
Pygments==2.3.1
PyHamcrest==2.0.2
pymitter==0.2.3
Pympler==0.6
PyNaCl==1.4.0
pyOpenSSL==20.0.0
pyotp==2.2.6
pyparsing==2.4.7
pyrsistent==0.17.3
pysha3==1.0.2
pyTelegramBotAPI==3.8.2
python-axolotl-curve25519==0.4.1.post2
python-binance==0.7.5
python-crontab==2.5.1
python-dateutil==2.7.5
python-editor==1.0.4
python-logstash==0.4.6
python3-openid==3.2.0
pytz==2018.9
PyWaves==0.8.36
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
repoze.lru==0.7
requests==2.26.0
requests-http-signature==0.2.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rlp==0.6.0
rsa==4.6
Rx==1.6.1
scrypt==0.8.6
service-identity==18.1.0
sha256==0.1
six==1.15.0
sniffio==1.2.0
solana==0.17.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
toolz==0.11.1
tqdm==4.29.1
translationstring==1.4
tronapi==3.1.6
trx-utils==1.0.4
Twisted==18.9.0
txaio==20.4.1
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
unicode==2.8
uritemplate==3.0.1
urllib3==1.24.1
varint==1.0.2
vine==5.0.0
virtualenv==16.2.0
wcwidth==0.1.7
web3==5.17.0
websockets==8.1
Werkzeug==1.0.1
zipp==3.4.0
zope.interface==5.2.0
